requirements.txt
Flask
pymongo
python-dotenv
certifi
gunicorn
streamlit
pandas
plotly